First pass at parsing disney scene resources. 64 bit file io functions because it's not 2012 any more.

+++ Source/Core/BuildCore/BuildDependencyGraph.cpp
@@ -62,7 +62,7 @@ namespace Selas
         }
 
         void* fileData;
-        uint32 fileSize;
+        uint64 fileSize;
         ReturnError_(File::ReadWholeFile(filepath.Ascii(), &fileData, &fileSize));
 
         CBinaryReadSerializer* serializer = New_(CBinaryReadSerializer);
