+++ Source/Core/IoLib/File.h
@@ -14,9 +14,9 @@ namespace Selas
     //=============================================================================================================================
     namespace File
     {
-        Error ReadWholeFile(cpointer filepath, void** fileData, uint32* fileSize);
-        Error ReadWhileFileAsString(cpointer filepath, char** string, uint32* stringSize);
-        Error WriteWholeFile(cpointer filepath, const void* data, uint32 size);
+        Error ReadWholeFile(cpointer filepath, void** fileData, uint64* fileSize);
+        Error ReadWhileFileAsString(cpointer filepath, char** string, uint64* stringSize);
+        Error WriteWholeFile(cpointer filepath, const void* data, uint64 size);
 
         bool Exists(cpointer filepath);
     };
