+++ Source/Core/SceneLib/SceneResource.cpp
@@ -94,7 +94,7 @@ namespace Selas
         AssetFileUtils::AssetFilePath(SceneResource::kDataType, SceneResource::kDataVersion, assetname, filepath);
 
         void* fileData = nullptr;
-        uint32 fileSize = 0;
+        uint64 fileSize = 0;
         ReturnError_(File::ReadWholeFile(filepath.Ascii(), &fileData, &fileSize));
 
         AttachToBinary(data->data, (uint8*)fileData, fileSize);
